mobot/requirements.txt
loguru==0.2.5
python-telegram-bot==12.0.0b1
requests
